Character concatination 0.2

+++ 0x06-pointers_arrays_strings/0-strcat.c
@@ -6,12 +6,11 @@
  * @date 2022-06-16
  * @copyright Copyright (c) 2022
  */
-#include <stdio.h>
 char *_strcat(char *dest, const char *src)
 {
-size_t dest_len = strlen(dest);
-size_t n = strlen(src);
-size_t i;
+int dest_len = strlen(dest);
+int n = strlen(src);
+int i;
 for (i = 0; i < n && src[i] != '\0'; i++)
 {
 dest[dest_len + i] = src[i];
